Abstract Crawly.Request via API
The Crawly.Request was a map (struct) before. Now we want to introduce
more logic to the requests, and one of the steps is to hide the
internal representation of the map by the API.

The new implementation provides a set of getters and setters to
operate with requests. We're not making any assumptions about the
the internal structure of Crawly.Request anymore.